Moved setGraphData method to own function in CreateGraphData.ts
